% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buccalbloodtensor.R
\docType{data}
\name{buccalbloodtensor}
\alias{buccalbloodtensor}
\title{The tested example data.}
\usage{
data(buccalbloodtensor)
}
\description{
The tested example data.
}
\references{
Teschendorff AE, Han J, Paul D, Virta J, Nordhausen K.
\emph{Tensorial Blind Source Separation for Improved Analysis of Multi-Omic Data.}
Genome Biology (2018) 19:76.
}
\keyword{datasets}
