% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_phenotype.R
\name{cor_phenotype}
\alias{cor_phenotype}
\title{Correlate inferred components against phenotype}
\usage{
cor_phenotype(tica.o, phenotype, phenotype.is.categorical, component = 0)
}
\arguments{
\item{tica.o}{the TICA result object, which can be obtained by function *DoTICA*.}

\item{phenotype}{A list comtains phenotype vectors inputed by users to correlate against, each should be in the same length with the number of sample.}

\item{phenotype.is.categorical}{A Boolean logical vector states whether the phenotype is categorical. It is of the same length of phenotype list.}

\item{component}{Choose the component significantly correlated to specific phenotype. It is an optional parameter for boxplot or scatter plot of chosen component against phenotype.}
}
\value{
A_star_matrix.p
A heatmap of the second mode mixing matrix.

pv.p
A heatmap of p value between phenotypes and the second mode mixing matrix components using linear regression model.

compheno.pl
A list of boxplots or scatter plots of chosen component against all phenotype.
}
\description{
Correlate inferred components against phenotype.
}
\examples{
data(buccalbloodtensor);
Dim.l <- EstDim(buccalbloodtensor$data);
dim <- Dim.l$dim;
tica.o <- DoTICA(Data = buccalbloodtensor$data, dim = dim, method = "FOBI");
phenotype.p <- cor_phenotype(tica.o = tica.o, phenotype = buccalbloodtensor$pheno.l,
 phenotype.is.categorical = buccalbloodtensor$pheno.i);
phenotype.p$A_star_matrix.p;
phenotype.p$pv.p;
 
}
\references{
Teschendorff AE, Han J, Paul D, Virta J, Nordhausen K.
\emph{Tensorial Blind Source Separation for Improved Analysis of Multi-Omic Data.}
Genome Biology (2018) 19:76.
}
