% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoTICA.R
\name{DoTICA}
\alias{DoTICA}
\title{Implement tensorial TICA}
\usage{
DoTICA(Data, dim, method = c("FOBI", "JADE"))
}
\arguments{
\item{Data}{The data object, which contains the multi-way data in two different formats. The data gives the array or data-tensor format. In the former case, and in our specific applications, the first mode defines the tissue, cell or data-type, the second mode defines the samples and the third mode the features (e.g. CpGs or genes). In the latter case, each list entry corresponds to the cell/tissue or data-type and consists of the data-matrix which rows representing features and columns representing samples.}

\item{dim}{A vector which contains the number of significant components of each data matrix to search for, and is typically obtained by applying RMT to each separate data/tissue-type matrix (i.e. to the individual entries of data$L above).}

\item{method}{Choose which TICA method,  tFOBI or tJADE.}
}
\value{
TICA.S
The source array result from TICA, which is of the same size as input data array containing the principal components.

TICA.projS
 List of projected TICA.S array to each sliced matrix with transforming back to the first mode.

TICA.W
	List containing all the unmixing matrices of TICA result.

TPCA.S
The source array result from TPCA, which is of the same size as input data array containing the principal components.

TPCA.U 
	List containing the rotation matrices of TPCA result.
}
\description{
Implement tensorial ICA(tFOBI and tJADE) to decompose the tensor-valued multi-omic data
}
\examples{
data(buccalbloodtensor);
Dim.l <- EstDim(buccalbloodtensor$data);
dim <- Dim.l$dim;
tica.o <- DoTICA(Data = buccalbloodtensor$data, dim = dim, method = "FOBI");

}
\references{
Teschendorff AE, Han J, Paul D, Virta J, Nordhausen K.
\emph{Tensorial Blind Source Separation for Improved Analysis of Multi-Omic Data.}
Genome Biology (2018) 19:76.

Virta, J., Li, B., Nordhausen, K. and Oja, H.
\emph{Independent component analysis for tensorvalued data}
Journal of Multivariate Analysis (2017).

Virta, J., Li, B., Nordhausen, K. and Oja, H.
\emph{JADE for Tensor-Valued Observation}
Journal of Computational and Graphical Statistics. preprint available on ArXiv
}
