% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDim.R
\name{EstDim}
\alias{EstDim}
\title{Estimate dim and dJ by RMT (Random Matrix Theory)}
\usage{
EstDim(Data)
}
\arguments{
\item{Data}{The array or data-tensor format of input data. In this case, and in our specific applications, the first mode defines the tissue, cell or data-type, the second mode defines the samples and the third mode the features (e.g. CpGs or genes).}
}
\value{
dim
a vector which contains the number of significant components of each data matrix to search for.

dJ
the number of significant components of joint variation across data/tissue types.
}
\description{
Auxiliary function to estimate dimensionality parameters
}
\examples{
data(buccalbloodtensor);
Dim.l <- EstDim(buccalbloodtensor$data);
dim <- Dim.l$dim;
dJ <- Dim.l$dJ;

}
\references{
Teschendorff AE, Han J, Paul D, Virta J, Nordhausen K.
\emph{Tensorial Blind Source Separation for Improved Analysis of Multi-Omic Data.}
Genome Biology (2018) 19:76.

Plerou, V., Gopikrishnan, P., Rosenow, B., Amaral, L.A., Guhr, T., Stanley, H.E. 
#' \emph{ Random matrix approach tocross correlations in financial data.} 
Phys Rev E Stat Nonlin Soft Matter Phys (2002) 65(6), 066126 

Teschendorff, A.E., Zhuang, J., Widschwendter, M.
\emph{Independent surrogate variable analysis to deconvolve confounding factors in large-scale microarray profiling studies.}
Bioinformatics (2011) 27(11), 1496-1505
}
