% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstSE.R
\name{EstSE}
\alias{EstSE}
\title{Estimate sensitivity}
\usage{
EstSE(tica.o, tp, topN = length(tp))
}
\arguments{
\item{tica.o}{The TICA result object, which can be obtained by function *DoTICA*.}

\item{tp}{An index vector labeling the true positive features associated with a factor of interest, which we know drives joint variation in the data, and which therefore the algorithm should capture.}

\item{topN}{The number of top-ranked features to select from each inferred component. It must be specified and by default it equals the number of true positives.}
}
\value{
se
The sensitivity list of all inferred components.

pv
The p value list of all inferred components.

se.p
A figure of 'se' list.

pv.p
A figure of 'pv' list.
}
\description{
Estimate sensitivity.
}
\examples{
data(buccalbloodtensor);
Dim.l <- EstDim(buccalbloodtensor$data);
dim <- Dim.l$dim;
tica.o <- DoTICA(Data = buccalbloodtensor$data, dim = dim, method = "FOBI");
SE <- EstSE(tica.o = tica.o, tp = seq(501,562));
SE$se.p;
SE$pv.p;
 
}
\references{
Teschendorff AE, Han J, Paul D, Virta J, Nordhausen K.
\emph{Tensorial Blind Source Separation for Improved Analysis of Multi-Omic Data.}
Genome Biology (2018) 19:76.
}
