% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_profile.R
\name{feature_profile}
\alias{feature_profile}
\title{Visualize the feature profile}
\usage{
feature_profile(
  Data,
  phenotype.v,
  phenotype.is.categorical,
  feature.index,
  tissue.index
)
}
\arguments{
\item{Data}{The data object used in TICA.}

\item{phenotype.v}{A vector comtains phenotypes inputed by users to plot against, it should be in the same length with the number of sample.}

\item{phenotype.is.categorical}{A Boolean logical variable states whether the phenotype is categorical.}

\item{feature.index}{The index of features in data matrix used in TICA. It can be obtained by feature selection.}

\item{tissue.index}{The index of tissue in data matrix used in TICA.}
}
\value{
profile.p
All the profile figures.
}
\description{
Visualize the feature profile against phenotype .
}
\examples{
data(buccalbloodtensor);
Dim.l <- EstDim(buccalbloodtensor$data);
dim <- Dim.l$dim;
tica.o <- DoTICA(Data = buccalbloodtensor$data, dim = dim, method = "FOBI");
feature.k <- feature_selection(tica.o = tica.o, component = 12, CLkurt = 0.95);
 (profile.p <- feature_profile(Data = buccalbloodtensor$data, 
phenotype.v = buccalbloodtensor$pheno.l$SmokingStatus, phenotype.is.categorical = TRUE,
feature.index = feature.k$pred.idx[[1]][seq_len(10)], tissue.index = 1))
}
\references{
Teschendorff AE, Han J, Paul D, Virta J, Nordhausen K.
\emph{Tensorial Blind Source Separation for Improved Analysis of Multi-Omic Data.}
Genome Biology (2018) 19:76.
}
