% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Feature selection}
\usage{
feature_selection(tica.o, component, topN = 0, CLkurt = 0)
}
\arguments{
\item{tica.o}{The TICA result object, which can be obtained by function *DoTICA*.}

\item{component}{The component ID.}

\item{topN}{The number of top-ranked features to select from each inferred component.}

\item{CLkurt}{The confidence level estimated by top-ranked features' kurtosis.}
}
\value{
pred.idx
The list of predicted feature index.

k.p
Visualizing the kurtosis of predicted feature.
}
\description{
Feature selection.
}
\examples{
data(buccalbloodtensor);
Dim.l <- EstDim(buccalbloodtensor$data);
dim <- Dim.l$dim;
tica.o <- DoTICA(Data = buccalbloodtensor$data, dim = dim, method = "FOBI");
feature.n <- feature_selection(tica.o = tica.o, component = 12, topN = 62);
buccalbloodtensor$testDMCs[feature.n$pred.idx[[1]]];
buccalbloodtensor$testDMCs[feature.n$pred.idx[[2]]];
feature.k <- feature_selection(tica.o = tica.o, component = 12, CLkurt = 0.95);
buccalbloodtensor$testDMCs[feature.k$pred.idx[[1]]];
buccalbloodtensor$testDMCs[feature.k$pred.idx[[2]]];
feature.k$k.p[[1]];
feature.k$k.p[[2]];
 
}
\references{
Teschendorff AE, Han J, Paul D, Virta J, Nordhausen K.
\emph{Tensorial Blind Source Separation for Improved Analysis of Multi-Omic Data.}
Genome Biology (2018) 19:76.
}
