% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_component.R
\name{plot_component}
\alias{plot_component}
\title{Visualize the weights of components}
\usage{
plot_component(tica.o, component, nt.idx, tp = 0)
}
\arguments{
\item{tica.o}{The TICA result object, which can be obtained by function *DoTICA*.}

\item{component}{The component ID to be plotted out.}

\item{nt.idx}{The first mode ID to be plotted out.}

\item{tp}{An index vector labeling the true positive features associated with a factor of interest, 
which we know drives joint variation in the data, and which therefore the algorithm should capture. 
These indices must be in the order of the entries in the 3rd mode of the data$A object, 
or alternatively in the same order as the rows of the individual data matrices in data$L.}
}
\value{
weight.p

weight_abs.p

weight_prime.p

weight_prime_abs.p
}
\description{
Visualize the weights of components to help with the feature selection.
}
\examples{
data(buccalbloodtensor);
Dim.l <- EstDim(buccalbloodtensor$data);
dim <- Dim.l$dim;
tica.o <- DoTICA(Data = buccalbloodtensor$data, dim = dim, method = "FOBI");
cp13 <- plot_component(tica.o = tica.o, component = 12, nt.idx = c(1,2), tp = seq(501,562));
cp13$weight.p;
cp13$weight_abs.p;
cp13$weight_prime.p;
cp13$weight_prime_abs.p;
 
}
\references{
Teschendorff AE, Han J, Paul D, Virta J, Nordhausen K.
\emph{Tensorial Blind Source Separation for Improved Analysis of Multi-Omic Data.}
Genome Biology (2018) 19:76.
}
